requirements.txt
transformers
deepspeed
sentencepiece
torch
protobuf
zmq
git+https://github.com/huggingface/transformers.git
wave
soundfile
scipy
accelerate
diffusers
llama-cpp-python
pygame
opencv-python
